Moved continous output and merge functions into data
